pyproject.toml
[project]
name = "anime-recommendation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langchain-groq>=1.0.1",
    "langchain-huggingface>=1.0.1",
    "langchain-pinecone>=0.2.13",
    "langgraph>=1.0.3",
    "langsmith>=0.4.45",
    "pandas>=2.3.3",
    "pinecone>=7.3.0",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "streamlit>=1.51.0",
]
